requirements.txt
captcha==0.3
Flask==1.1.2
Flask-Cors==3.0.9
peewee==3.14.0
Pillow==8.0.1
requests==2.25.0
psycopg2-binary~=2.8.6